@@ -1,7 +1,7 @@
 /*
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS HEADER.
  *
- * Copyright (c) 1997-2017 Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 1997-2021 Oracle and/or its affiliates. All rights reserved.
  *
  * The contents of this file are subject to the terms of either the GNU
  * General Public License Version 2 only ("GPL") or the Common Development
@@ -58,7 +58,7 @@ import java.nio.charset.*;
 import java.util.Iterator;
 import java.util.Set;
 
-import javax.xml.soap.*;
+import jakarta.xml.soap.*;
 
 public class SOAPListenerImpl implements SOAPListener {
 
