@@ -1,7 +1,7 @@
 /*
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS HEADER.
  *
- * Copyright (c) 1997-2017 Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 1997-2021 Oracle and/or its affiliates. All rights reserved.
  *
  * The contents of this file are subject to the terms of either the GNU
  * General Public License Version 2 only ("GPL") or the Common Development
@@ -43,8 +43,8 @@ package com.sun.xml.messaging.soap;
 import java.util.Iterator;
 import java.util.Vector;
 
-import javax.xml.soap.SOAPHeaderElement;
-import javax.xml.soap.SOAPException;
+import jakarta.xml.soap.SOAPHeaderElement;
+import jakarta.xml.soap.SOAPException;
 
 public abstract class SOAPRecipient {
 
